Smart-Recruits
Analytics Vidhya Hackathon

The Managers at 'Fintro' are primarily responsible for recruiting agents. Once a manager has identified a potential applicant, the would explain the business opportunity to the agent. Once the agent provides the consent, an application is made to Fintro to become an agent. This date is known as application_receipt_date. In the next 3 months, this potential agent has to undergo a 7 day training at the Fintro branch (about Sales processes and various products) and clear a subsequent examination in order to become a Fintro agent.

The problem - Who are the best agents? As is obvious in the above process, there is a significant investment which Fintro makes in identifying, training and recruiting these agents. However, there are a set of agents who do not bring in the expected resultant business. Fintro is looking for help from data scientists like you to help them provide insights using their past recruitment data. They want to predict the target variable for each potential agent, which would help them identify the right agents to hire.

Evaluation Metrics: ROC-AUC

Public LB Score: 0.864 Private LB Score: 0.760